+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All Notable changes to `isbn` will be documented in this file
 
+## 80% complete - 2015-02-19
+
+### Added
+Implemented:
+- `removeFormatting()`
+- `validate()`
+- `type()`
+- `is10()`
+- `is13()`
+- `hyphenate()`
+- `getRanges()`
+
+All except `hyphenate` and `getRanges` have full test coverage.
+Hyphenate() currently only works on ISBN13.
+
 ## Hello world - 2015-02-19
 
 ### Added